moved scoring type tests to individual modules
